[Socket Pool] Swap all int counters for size_t
We really should be using size_t for all socket capacity related numbers
as they should never be negative. I'm sure there are some spots I
missed, but this should be good enough to ensure forward propagation.

Some notes:
* I did re-work some SocketPoolAdditionalCapacityError constants, but
  it's not in use yet so whatever.
* I used uint32_t in network_service.mojom to be sure that size_t would
  equal or exceed it. I was worried that using uint64_t was a mistake as
  it would be bigger than size_t on 32bit builds. Also, we would never
  have more than uint32_t sockets anyway.
* I was able to remove a lot of non-negative CHECKs

Bug: 415691664
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7134007
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1544386}

+++ ios/chrome/browser/application_context/model/application_context_impl.mm
@@ -718,7 +718,7 @@ void ApplicationContextImpl::CreateLocalState() {
 
   net::ClientSocketPoolManager::set_max_sockets_per_proxy_chain(
       net::HttpNetworkSession::NORMAL_SOCKET_POOL,
-      std::max(std::min<int>(net::kDefaultMaxSocketsPerProxyChain, 99),
+      std::max(std::min<size_t>(net::kDefaultMaxSocketsPerProxyChain, 99u),
                net::ClientSocketPoolManager::max_sockets_per_group(
                    net::HttpNetworkSession::NORMAL_SOCKET_POOL)));
