+++ net/socket/client_socket_pool_manager.cc
@@ -33,7 +33,7 @@ namespace {
 // The limit for active sockets per pool for this network process. This may be
 // modified by set_max_sockets_per_pool in tests, but should otherwise be as
 // stated below.
-auto g_max_sockets_per_pool = std::to_array<int>({
+auto g_max_sockets_per_pool = std::to_array<size_t>({
     256,  // NORMAL_SOCKET_POOL
     256   // WEBSOCKET_SOCKET_POOL
 });
@@ -50,7 +50,7 @@ static_assert(std::size(g_max_sockets_per_pool) ==
 // than normal other connections. Use a limit of 255, so the limit for wss will
 // be the same as the limit for ws. Also note that Firefox uses a limit of 200.
 // See http://crbug.com/486800
-auto g_max_sockets_per_group = std::to_array<int>({
+auto g_max_sockets_per_group = std::to_array<size_t>({
     6,   // NORMAL_SOCKET_POOL
     255  // WEBSOCKET_SOCKET_POOL
 });
@@ -62,7 +62,7 @@ static_assert(std::size(g_max_sockets_per_group) ==
 // The max number of sockets to allow per proxy chain.  This applies both to
 // http and SOCKS proxies.  See http://crbug.com/12066 and
 // http://crbug.com/44501 for details about proxy chain connection limits.
-auto g_max_sockets_per_proxy_chain = std::to_array<int>({
+auto g_max_sockets_per_proxy_chain = std::to_array<size_t>({
     kDefaultMaxSocketsPerProxyChain,  // NORMAL_SOCKET_POOL
     kDefaultMaxSocketsPerProxyChain   // WEBSOCKET_SOCKET_POOL
 });
@@ -143,7 +143,7 @@ ClientSocketPoolManager::ClientSocketPoolManager() = default;
 ClientSocketPoolManager::~ClientSocketPoolManager() = default;
 
 // static
-int ClientSocketPoolManager::max_sockets_per_pool(
+size_t ClientSocketPoolManager::max_sockets_per_pool(
     HttpNetworkSession::SocketPoolType pool_type) {
   DCHECK_LT(pool_type, HttpNetworkSession::NUM_SOCKET_POOL_TYPES);
   return g_max_sockets_per_pool[pool_type];
@@ -152,9 +152,9 @@ int ClientSocketPoolManager::max_sockets_per_pool(
 // static
 void ClientSocketPoolManager::set_max_sockets_per_pool_for_test(
     HttpNetworkSession::SocketPoolType pool_type,
-    int socket_count) {
-  DCHECK_LT(0, socket_count);     // At least one socket must be allowed.
-  DCHECK_GE(2048, socket_count);  // For now, we pick a ceiling of 2^11.
+    size_t socket_count) {
+  DCHECK_LT(0u, socket_count);     // At least one socket must be allowed.
+  DCHECK_GE(2048u, socket_count);  // For now, we pick a ceiling of 2^11.
   DCHECK_LT(pool_type, HttpNetworkSession::NUM_SOCKET_POOL_TYPES);
   g_max_sockets_per_pool[pool_type] = socket_count;
   DCHECK_GE(g_max_sockets_per_pool[pool_type],
@@ -162,7 +162,7 @@ void ClientSocketPoolManager::set_max_sockets_per_pool_for_test(
 }
 
 // static
-int ClientSocketPoolManager::max_sockets_per_group(
+size_t ClientSocketPoolManager::max_sockets_per_group(
     HttpNetworkSession::SocketPoolType pool_type) {
   DCHECK_LT(pool_type, HttpNetworkSession::NUM_SOCKET_POOL_TYPES);
   return g_max_sockets_per_group[pool_type];
@@ -171,9 +171,9 @@ int ClientSocketPoolManager::max_sockets_per_group(
 // static
 void ClientSocketPoolManager::set_max_sockets_per_group_for_test(
     HttpNetworkSession::SocketPoolType pool_type,
-    int socket_count) {
-  DCHECK_LT(0, socket_count);    // At least one socket must be allowed.
-  DCHECK_GE(512, socket_count);  // For now, we pick a ceiling of 2^9.
+    size_t socket_count) {
+  DCHECK_LT(0u, socket_count);    // At least one socket must be allowed.
+  DCHECK_GE(512u, socket_count);  // For now, we pick a ceiling of 2^9.
   DCHECK_LT(pool_type, HttpNetworkSession::NUM_SOCKET_POOL_TYPES);
   g_max_sockets_per_group[pool_type] = socket_count;
 
@@ -184,7 +184,7 @@ void ClientSocketPoolManager::set_max_sockets_per_group_for_test(
 }
 
 // static
-int ClientSocketPoolManager::max_sockets_per_proxy_chain(
+size_t ClientSocketPoolManager::max_sockets_per_proxy_chain(
     HttpNetworkSession::SocketPoolType pool_type) {
   DCHECK_LT(pool_type, HttpNetworkSession::NUM_SOCKET_POOL_TYPES);
   return g_max_sockets_per_proxy_chain[pool_type];
@@ -193,9 +193,9 @@ int ClientSocketPoolManager::max_sockets_per_proxy_chain(
 // static
 void ClientSocketPoolManager::set_max_sockets_per_proxy_chain(
     HttpNetworkSession::SocketPoolType pool_type,
-    int socket_count) {
-  DCHECK_LT(0, socket_count);    // At least one socket must be allowed.
-  DCHECK_GE(128, socket_count);  // For now, we pick a ceiling of 2^7.
+    size_t socket_count) {
+  DCHECK_LT(0u, socket_count);    // At least one socket must be allowed.
+  DCHECK_GE(128u, socket_count);  // For now, we pick a ceiling of 2^7.
   DCHECK_LT(pool_type, HttpNetworkSession::NUM_SOCKET_POOL_TYPES);
   // Assert this case early on. The max number of sockets per group cannot
   // exceed the max number of sockets per proxy chain.
